fix: Change by extending ResponseEntityExceptionHandler
